preprocessing/requirements.txt
cleantext
numpy
pandas
torch
transformers
sentencepiece
datasets
black
python-dotenv
openpyxl
sentence_transformers
nltk